Restructuring login and signup
Restructured the log in and sign up pages so that they are in the "Pages" folder
